Feat: User Auth with JWT.

+++ resources/item.py
@@ -1,6 +1,6 @@
-from flask import request
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
+from flask_jwt_extended import jwt_required, get_jwt
 from sqlalchemy.exc import SQLAlchemyError
 
 from db import db
@@ -9,14 +9,21 @@ from schemas import ItemSchema, ItemUpdateSchema
 
 blp = Blueprint("Items", __name__, description="Operation on items.")
 
-@blp.route("/item/<string:itemID>")
+@blp.route("/item/<int:itemID>")
 class Item(MethodView):
+    @jwt_required()
     @blp.response(200, ItemSchema)
-    def get(self, itemID):
+    def get(self, itemID:int) -> dict:
         item = ItemModel.query.get_or_404(itemID)
         return item
+    
+    @jwt_required()
+    def delete(self, itemID:int) -> dict:
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401,
+                  message="Admin privilege required.")
             
-    def delete(self, itemID):
         item = ItemModel.query.get_or_404(itemID)
         db.session.delete(item)
         db.session.commit()
@@ -25,7 +32,7 @@ class Item(MethodView):
     # Update item
     @blp.arguments(ItemUpdateSchema)
     @blp.response(200, ItemSchema)
-    def put(self, itemData, itemID):
+    def put(self, itemData:dict, itemID:int) -> dict:
         item = ItemModel.query.get(itemID)
         if item:
             item.price = itemData["price"]
@@ -40,13 +47,15 @@ class Item(MethodView):
     
 @blp.route("/item")
 class ItemList(MethodView):
+    @jwt_required()
     @blp.response(200, ItemSchema(many=True))
-    def get(self):
+    def get(self) -> dict:
         return ItemModel.query.all()
     
+    @jwt_required(fresh=True)
     @blp.arguments(ItemSchema)    
     @blp.response(201, ItemSchema)
-    def post(self, itemData):
+    def post(self, itemData:dict) -> dict:
         item = ItemModel(**itemData)
         
         try:
@@ -55,5 +64,5 @@ class ItemList(MethodView):
         except SQLAlchemyError:
             abort(500, message="An error occurred while inserting items.")
         
-        
+
         return item, 201
